apps/data/pyproject.toml
[project]
name = "versetools-data"
version = "0.0.0"
requires-python = ">=3.10, <3.12"
dependencies = [
    "convex~=0.7.0",
    "python-dotenv~=1.2.2",
    "fnvhash~=0.1.0",
    "hexdump~=3.3",
    "humanize~=4.11.0",
    "numpy==1.24.3",
    "packaging~=24.1",
    "pillow~=10.4.0",
    "pycryptodome~=3.21.0",
    "pynubia~=0.3.0",
    "pyquaternion~=0.9.9",
    "pyrsi~=0.1.19",
    "rich~=13.9.4",
    "sentry-sdk~=1.5.8",
    "tqdm~=4.66.6",
    "xxhash~=3.5.0",
    "zstandard~=0.23.0",
]

[dependency-groups]
dev = [
    "ipython>=8.29.0",
    "ruff>=0.7.2",
]
